refactor: use ethers conditional exports
re #632
