+++ packages/data/tsconfig.json
@@ -1,4 +1,4 @@
 {
     "extends": "../../tsconfig.json",
-    "include": ["src", "rollup.config.ts"]
+    "include": ["src", "tests", "rollup.config.ts"]
 }
